requirements.txt
# Core dependencies for testLLM framework
pytest>=7.0.0
PyYAML>=6.0
requests>=2.25.0
python-dotenv>=0.19.0

# Optional development dependencies
# Install with: pip install -e ".[dev]"